+++ src/ui/window.py
@@ -207,7 +207,7 @@ class MainWindow(Adw.ApplicationWindow):
         about.set_application_name(self.lang_manager.get("window.about.name"))
         about.set_application_icon("io.github.askscience.gaia")
         about.set_developer_name(self.lang_manager.get("window.about.developer"))
-        about.set_version("0.3.0")
+        about.set_version("0.3.1")
         about.set_comments(self.lang_manager.get("window.about.comments"))
         about.set_copyright(self.lang_manager.get("window.about.copyright"))
         about.set_website("https://github.com/askscience/gaia")